Devices: Drop Xiaomi Redmi Note 10 (Mojito)
Tejas Kamat has been proven to be in possession of child pornography, as we take the safety of the public seriously, his maintainership has been previously officially terminated as of Tuesday, June 14th, 2022 - 16:29 Japan Standard Time. We have since worked with several people in the community to ensure that he is unable to commit any further threats to society.

Since then, this particular individual does not have any involvement with us, any actions he may claim to do on behalf of our project before or after the termination are false and should not be trusted.

Should you have any concerns with our maintainer, reach out to trust@evolution-x.org.

Change-Id: I39f3d416c5096923d289a2962337c75fafeab6ee
Signed-off-by: Akito Mizukito <akito@evolution-x.org>