[EvolutionX-CI]: Devices: Add new device_group property
This commit adds a new device_group property to devices.json that allows
maintainers to specify the telegram group for a given device which can be
updated when creating pull requests for OTAs if needed. This not only
simplifies the generation script, but also allows us to easily add
the groups to Evolution-X.org in future.

The device_group property is subject to the following requirements:
- It must be Evolution X specific.
- It cannot be a channel.
- The device group must have the "EvolutionX" prefix.

Signed-off-by: Akito Mizukito <akito@evolution-x.org>