generate_device_groups: Add missing dependencies check
- Checks if the required dependencies are available. If missing, it prompts
the user for confirmation before proceeding to install the missing dependencies while
providing a description of each dependency that will be installed. If the user declines,
the script is halted.

- Checks the availability of apt-get, pacman, and dnf using
the command -v utility. Depending on which command is found, the script
determines the corresponding Linux distribution (Debian/Ubuntu, Arch Linux,or Fedora)
to proceed with.

Signed-off-by: AnierinB <anierin@evolution-x.org>